feat(collection): stabilize dropLastWhile
