+++ collections/drop_last_while_test.ts
@@ -10,11 +10,9 @@ Deno.test("dropLastWhile() handles num array", () => {
   assertEquals(actual, [20]);
 });
 
-Deno.test("dropLastWhile() handles no mutation", () => {
+Deno.test("dropLastWhile() does not mutate the input array", () => {
   const array = [1, 2, 3, 4, 5, 6];
-
   const actual = dropLastWhile(array, (i) => i > 4);
-
   assertEquals(actual, [1, 2, 3, 4]);
   assertEquals(array, [1, 2, 3, 4, 5, 6]);
 });
@@ -49,4 +47,36 @@ Deno.test("dropLastWhile() returns empty array when all elements get dropped", (
   const actual = dropLastWhile(array, (i) => i < 40);
 
   assertEquals(actual, []);
+});
+
+Deno.test("dropLastWhile() handles a string", () => {
+  const values = "hello there world";
+  const actual = dropLastWhile(values, (i) => i !== " ");
+  assertEquals(actual, "hello there ".split(""));
+});
+
+Deno.test("dropLastWhile() handles a Set", () => {
+  const values = new Set([20, 33, 44]);
+  const actual = dropLastWhile(values, (i) => i > 30);
+  assertEquals(actual, [20]);
+});
+
+Deno.test("dropLastWhile() handles a Map", () => {
+  const values = new Map([
+    ["a", 20],
+    ["b", 33],
+    ["c", 44],
+  ]);
+  const actual = dropLastWhile(values, ([_k, v]) => v > 30);
+  assertEquals(actual, [["a", 20]]);
+});
+
+Deno.test("dropLastWhile() handles a generator", () => {
+  function* gen() {
+    yield 20;
+    yield 33;
+    yield 44;
+  }
+  const actual = dropLastWhile(gen(), (i) => i > 30);
+  assertEquals(actual, [20]);
 });
